Change tests to be based on JUnit, rename comments variable
